+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - (Feature) (ML) Container Envs
 - (Feature) Pass Arguments to ID Group
 - (Documentation) (ML) Minimal docs with examples how to deploy ArangoMLExtension
+- (Feature) Add --deployment.feature.init-containers-upscale-resources (default enabled)
 
 ## [1.2.35](https://github.com/arangodb/kube-arangodb/tree/1.2.35) (2023-11-06)
 - (Maintenance) Update go-driver to v1.6.0, update IsNotFound() checks